refactor: move home screen components to shared components
