+++ app/src/main/java/com/amsterdam/cutetudee/presentation/component/sharedComponent/ImageMood.kt
@@ -1,4 +1,4 @@
-package com.amsterdam.cutetudee.presentation.screens.home.component
+package com.amsterdam.cutetudee.presentation.component.sharedComponent
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
@@ -11,6 +11,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.painter.Painter
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.unit.dp
+import com.amsterdam.cutetudee.R
 import com.amsterdam.cutetudee.presentation.theme.AppTheme
 import com.amsterdam.cutetudee.presentation.utils.ThemeAndLocalePreviews
 
@@ -36,6 +37,6 @@ fun ImageMood(
 @ThemeAndLocalePreviews
 private fun ImageMoodPreview() {
     ImageMood(
-        image = painterResource(id = com.amsterdam.cutetudee.R.drawable.tudee_image_neutral),
+        image = painterResource(id = R.drawable.tudee_image_neutral),
     )
 }
